Fix formatting along with other issues

+++ src/lib.rs
@@ -42,12 +42,15 @@ impl Span {
     }
 
     /// Grows the span from the back. This moves the start value back by `amount`.
-    /// 
+    ///
     /// # Panics
     /// Panics if the start of the span is less than `amount`, since spans can't have a negative start value,
     #[inline(always)]
     pub fn grow_back(&mut self, amount: SpanValue) {
-        assert!(self.start >= amount, "cannot create a span with a negative start value");
+        assert!(
+            self.start >= amount,
+            "cannot create a span with a negative start value"
+        );
         self.start -= amount;
     }
 
@@ -56,8 +59,12 @@ impl Span {
     /// # Panics
     /// Panics if the size of the `Span` is less than `amount`, since a `Span`'s size can't be negative.
     #[inline(always)]
+    #[allow(clippy::unnecessary_cast)]
     pub fn shrink_back(&mut self, amount: SpanValue) {
-        assert!(self.len() >= amount, "cannot create negative-size span");
+        assert!(
+            self.len() >= amount as usize,
+            "cannot create negative-size span"
+        );
         self.start += amount;
     }
 
@@ -66,8 +73,12 @@ impl Span {
     /// # Panics
     /// This method will panic if the size of the `Span` is less than `amount`, since a `Span`'s size can't be negative.
     #[inline(always)]
+    #[allow(clippy::unnecessary_cast)]
     pub fn shrink_front(&mut self, amount: SpanValue) {
-        assert!(self.len() >= amount, "cannot create negative-size span");
+        assert!(
+            self.len() >= amount as usize,
+            "cannot create negative-size span"
+        );
         self.end -= amount;
     }
 
@@ -93,12 +104,15 @@ impl Span {
     }
 
     /// Applies the span to `string`.
-    /// 
+    ///
     /// # Panics
     /// Panics if `string` is shorter than the end of the span.
     #[allow(clippy::unnecessary_cast)]
     pub fn apply<'a>(&self, string: &'a str) -> &'a str {
-        assert!(string.len() > self.end, "string is too short to have the span applied");
+        assert!(
+            string.len() > self.end as usize,
+            "string is too short to have the span applied"
+        );
         let start = string.char_indices().nth(self.start as usize).unwrap().0;
         let end = string.char_indices().nth(self.end as usize).unwrap().0;
         &string[start..end]
